only setup environment if vagrant is installed

+++ bash.d/72-vagrant.sh
@@ -1,4 +1,6 @@
-alias v='vagrant'
-alias vd='vagrant destroy'
-alias vu='vagrant up'
-alias vs='vagrant ssh'
+if [[ $(:binexists vagrant) -eq 0 ]]; then
+  alias v='vagrant'
+  alias vd='vagrant destroy'
+  alias vu='vagrant up'
+  alias vs='vagrant ssh'
+fi
